requirements.txt
pandas
numpy
scikit-learn
wwo-hist
python-dotenv
requests